@86eqxm068 #Moved date data on block
