--- apps/web-customer/src/components/blocks/ComingSoonBlock/ComingSoonData.ts
@@ -1,6 +1,0 @@
-export const dateData = [{
-  days: "Days",
-  hours: "Hours",
-  minutes: "Minutes",
-  seconds: "Seconds",
-}]
